Move detail into feature.detail
