+++ feature/detail/src/main/java/ru/d3rvich/feature/detail/model/GameDetailUiEvent.kt
@@ -1,4 +1,4 @@
-package ru.d3rvich.detail.model
+package ru.d3rvich.feature.detail.model
 
 import ru.d3rvich.core.ui.base.UiEvent
