+++ feature/detail/src/main/java/ru/d3rvich/feature/detail/model/GameDetailUiModel.kt
@@ -1,4 +1,4 @@
-package ru.d3rvich.detail.model
+package ru.d3rvich.feature.detail.model
 
 import androidx.compose.runtime.Immutable
 import kotlinx.collections.immutable.ImmutableList
