+++ feature/detail/src/main/java/ru/d3rvich/feature/detail/views/GameDetailView.kt
@@ -1,4 +1,4 @@
-package ru.d3rvich.detail.views
+package ru.d3rvich.feature.detail.views
 
 import androidx.compose.foundation.border
 import androidx.compose.foundation.horizontalScroll
@@ -56,9 +56,9 @@ import ru.d3rvich.core.ui.icon.tryFindIcon
 import ru.d3rvich.core.ui.theme.DarkBlue
 import ru.d3rvich.core.ui.theme.Purple
 import ru.d3rvich.detail.R
-import ru.d3rvich.detail.model.GameDetailUiModel
-import ru.d3rvich.detail.model.ScreenshotsUiState
-import ru.d3rvich.detail.model.toGameDetailUiModel
+import ru.d3rvich.feature.detail.model.GameDetailUiModel
+import ru.d3rvich.feature.detail.model.ScreenshotsUiState
+import ru.d3rvich.feature.detail.model.toGameDetailUiModel
 
 /**
  * Created by Ilya Deryabin at 11.03.2024
